+++ bespokeasm/assembler/model/operand/factory.py
@@ -4,12 +4,12 @@
 
 
 from bespokeasm.assembler.model.operand import Operand
-from bespokeasm.assembler.model.operand.types import empty, numeric_expression, indirect_register, indirect_numeric, register
+from bespokeasm.assembler.model.operand.types import empty, numeric_expression, indirect_register, indirect_numeric, register, indirect_indexed_register
 
 class OperandFactory:
 
     @classmethod
-    def factory(cls, operand_id: str, arg_config_dict: dict, default_endian: str) -> Operand:
+    def factory(cls, operand_id: str, arg_config_dict: dict, default_endian: str, ) -> Operand:
         type_str = arg_config_dict['type']
         if type_str == 'numeric':
             return numeric_expression.NumericExpressionOperand(operand_id, arg_config_dict, default_endian)
@@ -17,6 +17,8 @@ class OperandFactory:
             return register.RegisterOperand(operand_id, arg_config_dict, default_endian)
         elif type_str == 'indirect_register':
             return indirect_register.IndirectRegisterOperand(operand_id, arg_config_dict, default_endian)
+        elif type_str == 'indirect_indexed_register':
+            return indirect_indexed_register.IndirectIndexedRegisterOperand(operand_id, arg_config_dict, default_endian)
         elif type_str == 'indirect_numeric':
             return indirect_numeric.IndirectNumericOperand(operand_id, arg_config_dict, default_endian)
         elif type_str == 'empty':